+++ src/imio/updates/update_instances.py
@@ -152,38 +152,46 @@ def get_plone_site(path):
     return name
 
 
-def get_instance_port(path):
-    line = None
-    cmd = 'grep instance1-http %s/port.cfg|cut -d" " -f 3' % path
-    (out, err, code) = runCommand(cmd)
-    for line in out:
-        line = line.strip('\n')
-        try:
-            int(line)
-        except ValueError:
-            continue
-        break
-    else:
-        error("Cannot extract instance1-http from '%s/port.cfg'" % path)
-    return line
+def get_instance_port(path, instance='instance1'):
+    proc_http_name= "%s-http" % instance
+    cmd = 'grep %s %s/port.cfg | cut -d" " -f 3' % (proc_http_name, path)
+    with open("%s/port.cfg" % path, mode='r') as file:
+        for line in file:
+            if proc_http_name in line:
+                port = line.split(' ')[-1].strip()
+                try:
+                    int(port)
+                    return port
+                except ValueError:
+                    continue
+    error("Cannot extract %s from '%s/port.cfg'" % (proc_http_name, path))
+    return None
 
 
-def run_spv(bldt, command, processes, wait=False):
+def run_spv(bldt, path, plone_path, command, processes, wait=False):
     for proc in processes:
-        cmd = 'supervisorctl %s %s-%s' % (command, bldt['bldt'], proc)
+        cmd = 'supervisorctl %s %s-%s' % (command, bldt, proc)
         if doit:
             verbose("=> Running '%s'" % cmd)
             (out, err, code) = runCommand(cmd)
             if code:
                 error("Problem running supervisor command")
-            elif wait and len(processes) > 1:
+            elif wait and proc != processes[-1]:
+                threshold = 20
+                interval = 5
+                verbose('Waiting %d sec ...' % threshold)
                 time.sleep(20)
+                port = get_instance_port(path, proc)
+                url = 'http://localhost:%s/%s/ok' % (port, plone_path)
                 for i in range(0, 9):
                     try:
-                        response = requests.get('http://localhost:%s/%s/ok' % (bldt['port'], bldt['plone']))
+                        verbose('Checking %s' % url)
+                        response = requests.get(url)
                         if response.status_code == 200:
                             break
                         else:
+                            verbose('Status HTTP status code was %d. Waiting another %d sec...' % (response.status_code,
+                                                                                                   interval))
                             time.sleep(5)
                     except Exception as err:
                         # Don't care the nature of this error
@@ -402,19 +410,19 @@ def main():
             error("Path '%s' doesn't exist" % path)
             continue
 
-        buildouts[bldt]['bldt'] = bldt
         buildouts[bldt]['path'] = path
-        buildouts[bldt]['plone'] = get_plone_site(path)
+        plone_path = get_plone_site(path)
+        buildouts[bldt]['plone'] = plone_path
         buildouts[bldt]['port'] = get_instance_port(path)
 
         verbose("Buildout %s" % path)
         if stop:
             if 'i' in stop:
-                run_spv(buildouts[bldt], 'stop', reversed([p for p in buildouts[bldt]['spv'] if p.startswith('instance')]))
+                run_spv(bldt, path, plone_path, 'stop', reversed([p for p in buildouts[bldt]['spv'] if p.startswith('instance')]))
             if 'a' in stop:
-                run_spv(buildouts[bldt], 'stop', reversed([p for p in buildouts[bldt]['spv']]))
+                run_spv(bldt, path, plone_path, 'stop', reversed([p for p in buildouts[bldt]['spv']]))
             if 'w' in stop:
-                run_spv(buildouts[bldt], 'stop', reversed([p for p in buildouts[bldt]['spv'] if p.startswith('worker')]))
+                run_spv(bldt, path, plone_path, 'stop', reversed([p for p in buildouts[bldt]['spv'] if p.startswith('worker')]))
 
         if ns.make0:
             for param_list in ns.make0:
@@ -427,11 +435,11 @@ def main():
             run_develop(buildouts, bldt, ns.develop)
         if restart:
             if 'i' in restart:
-                run_spv(buildouts[bldt], 'restart', [p for p in buildouts[bldt]['spv'] if p.startswith('instance')], wait=True)
+                run_spv(bldt, path, plone_path, 'restart', [p for p in buildouts[bldt]['spv'] if p.startswith('instance')], wait=True)
             if 'a' in restart:
-                run_spv(buildouts[bldt], 'restart', [p for p in buildouts[bldt]['spv']])
+                run_spv(bldt, path, plone_path, 'restart', [p for p in buildouts[bldt]['spv']])
             if 'w' in restart:
-                run_spv(buildouts[bldt], 'restart', [p for p in buildouts[bldt]['spv'] if p.startswith('worker')])
+                run_spv(bldt, path, plone_path, 'restart', [p for p in buildouts[bldt]['spv'] if p.startswith('worker')])
 
         if 'zeoserver' not in buildouts[bldt]['spv']:
             error("Zeoserver isn't running")